refactor(review): ReviewEvalPanel を3つのサブコンポーネントに分割
レビュー困難な巨大 SFC を責務別に分離:
- ReviewVerdict.vue: 結論ヘッダー＋スコア（および空・解析中・CPU 手のフォールバック）
- ReviewCandidateGrid.vue: 候補手グリッド
- ReviewProgression.vue: 進行ツリー（最善・実際・被詰タブ + インライン分岐）
- composables/useReviewProgression.ts: 進行ツリーの状態と行構築ロジック
- ReviewEvalPanel.vue: オーケストレータに縮小（96 行）

外部 API（props/emits）は据え置き、CpuReviewPlayer.vue 側は変更不要。
